// backend/actions/event/types.ts
export enum EventState {
    UNKNOWN = 'UNKNOWN',

    EVENT_GROUPING = 'EVENT_GROUPING',

    EVENT_TRANSMISSION_SCREEN = 'EVENT_TRANSMISSION_SCREEN',
    EVENT_LOADING_SCREEN = 'EVENT_LOADING_SCREEN',

    EVENT_INSTANCE_IN_PROGRESS = 'EVENT_INSTANCE_IN_PROGRESS',
    EVENT_INSTANCE_FINISHED = 'EVENT_INSTANCE_FINISHED',
    EVENT_INSTANCE_EXIT = 'EVENT_INSTANCE_EXIT',

    EVENT_INSTANCE_EXITED = 'EVENT_INSTANCE_EXITED',
}